% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalCmprsk-functions.R
\name{fit.nonpar}
\alias{fit.nonpar}
\title{Nonparametric estimation of ATE corresponding to the target population}
\usage{
fit.nonpar(
  df,
  T,
  E,
  A,
  C,
  wtype = "stab.ATE",
  bs = FALSE,
  nbs.rep = 400,
  seed = 17,
  cens = 0,
  conf.level = 0.95
)
}
\arguments{
\item{df}{a data frame with time to event, type of event, treatment indicator and covariates.}

\item{T}{a character string specifying the name of the time-to-event variable in \code{df}.}

\item{E}{a character string specifying the name of the "event type" variable in \code{df}.}

\item{A}{a character specifying the name of the treatment/exposure variable.
It is assumed that \code{A} is a numeric binary indicator with 0/1 values, where \code{A}=1
is assumed a treatment group, and \code{A}=0 a control group.}

\item{C}{a vector of character strings with variable names (potential confounders)
in the logistic regression model for Propensity Scores, i.e. P(A=1|C=c).}

\item{wtype}{a character variable indicating the type of weight.
The default is "stab.ATE" defined as w_stab=P(A=a)/P(A=a|C=c) - see Hernan et al. (2000).
Other possible values are "ATE", "ATT", "ATC" and "overlap". See Table 1 from Li, Morgan, and Zaslavsky (2018).
There is also an option of "unajusted" estimation, by settinh wtype="unadj" - this will not adjust for possible
treatment selection bias and will not use propensity scores weighting. It can be used, for example,
in data from an RCT where there is no need for emulation of baseline randomization.}

\item{seed}{for the bootstrap}

\item{cens}{integer value in \code{E} that corresponds to censorings resorded in \code{T}.}
}
\value{
A list with components:
}
\description{
Description ...
}
\examples{
# please see our package vignette for practical examples

}
