% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalCmprsk-functions.R
\name{get.pointEst}
\alias{get.pointEst}
\title{Returns point estimates corresponding to a time point}
\usage{
get.pointEst(cmprsk.obj, timepoint)
}
\arguments{
\item{cmprsk.obj}{a "cmprsk" object returned by one of the function \code{fit.cox} or \code{fit.nonpar}}

\item{timepoint}{a scalar value of the time point of interest}
}
\description{
Returns point estimates corresponding to a time point
}
