% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalCmprsk-functions.R
\name{get.weights}
\alias{get.weights}
\title{Fitting propensity scores model and estimating weights}
\usage{
get.weights(df, A, C, wtype = "stab.ATE", case.w = NULL)
}
\arguments{
\item{case.w}{a vector of sampling weights}
}
\description{
Description...
}
\references{
F. Li, K.L. Morgan, and A.M. Zaslavsky. 2018. Balancing Covariates via Propensity Score Weighting. Journal of the American Statistical Association 113 (521): 390–400.

M. Hernan, K.L. Morgan, and A.M. Zaslavsky. 2000. Balancing Covariates via Propensity Score Weighting. Journal of the American Statistical Association 113 (521): 390–400.
}
